multiboot2-header: moved all related code into the new `builder`-module
