+++ multiboot2-header/src/builder/mod.rs
@@ -1,0 +1,8 @@
+//! Module for the builder-feature.
+
+mod header;
+mod information_request;
+pub(self) mod traits;
+
+pub use header::Multiboot2HeaderBuilder;
+pub use information_request::InformationRequestHeaderTagBuilder;
