+++ multiboot2-header/src/entry_header.rs
@@ -49,6 +49,3 @@ impl Debug for EntryHeaderTag {
             .finish()
     }
 }
-
-#[cfg(feature = "builder")]
-impl crate::StructAsBytes for EntryHeaderTag {}
