+++ multiboot2-header/src/information_request.rs
@@ -1,11 +1,5 @@
 use crate::HeaderTagType;
-#[cfg(feature = "builder")]
-use crate::StructAsBytes;
 use crate::{HeaderTagFlag, MbiTagType};
-#[cfg(feature = "builder")]
-use alloc::collections::BTreeSet;
-#[cfg(feature = "builder")]
-use alloc::vec::Vec;
 use core::fmt;
 use core::fmt::{Debug, Formatter};
 use core::marker::PhantomData;
@@ -92,87 +86,6 @@ impl<const N: usize> Debug for InformationRequestHeaderTag<N> {
     }
 }
 
-#[cfg(feature = "builder")]
-impl<const N: usize> crate::StructAsBytes for InformationRequestHeaderTag<N> {}
-
-/// Helper to build the dynamically sized [`InformationRequestHeaderTag`]
-/// at runtime.
-#[derive(Debug)]
-#[cfg(feature = "builder")]
-pub struct InformationRequestHeaderTagBuilder {
-    flag: HeaderTagFlag,
-    irs: BTreeSet<MbiTagType>,
-}
-
-#[cfg(feature = "builder")]
-impl InformationRequestHeaderTagBuilder {
-    /// New builder.
-    pub fn new(flag: HeaderTagFlag) -> Self {
-        Self {
-            irs: BTreeSet::new(),
-            flag,
-        }
-    }
-
-    /// Returns the expected length of the information request tag,
-    /// when the `build`-method gets called.
-    pub fn expected_len(&self) -> usize {
-        let basic_header_size = size_of::<InformationRequestHeaderTag<0>>();
-        let req_tags_size = self.irs.len() * size_of::<MbiTagType>();
-        basic_header_size + req_tags_size
-    }
-
-    /// Adds an [`MbiTagType`] to the information request.
-    pub fn add_ir(mut self, tag: MbiTagType) -> Self {
-        self.irs.insert(tag);
-        self
-    }
-
-    /// Adds multiple [`MbiTagType`] to the information request.
-    pub fn add_irs(mut self, tags: &[MbiTagType]) -> Self {
-        self.irs.extend(tags);
-        self
-    }
-
-    /// Builds the bytes of the dynamically sized information request header.
-    pub fn build(self) -> Vec<u8> {
-        let expected_len = self.expected_len();
-        let mut data = Vec::with_capacity(expected_len);
-
-        let basic_tag = InformationRequestHeaderTag::<0>::new(
-            self.flag,
-            [],
-            // we put the expected length here already, because in the next step we write
-            // all the tags into the byte array. We can't know this during compile time,
-            // therefore N is 0.
-            Some(expected_len as u32),
-        );
-        data.extend(basic_tag.struct_as_bytes());
-        #[cfg(debug_assertions)]
-        {
-            let basic_tag_size = size_of::<InformationRequestHeaderTag<0>>();
-            assert_eq!(
-                data.len(),
-                basic_tag_size,
-                "the vector must be as long as the basic tag!"
-            );
-        }
-
-        for tag in &self.irs {
-            let bytes: [u8; 4] = (*tag as u32).to_ne_bytes();
-            data.extend(&bytes);
-        }
-
-        debug_assert_eq!(
-            data.len(),
-            expected_len,
-            "the byte vector must be as long as the expected size of the struct"
-        );
-
-        data
-    }
-}
-
 /// Iterates the dynamically sized information request structure and finds all MBI tags
 /// that are requested.
 #[derive(Copy, Clone)]
@@ -215,38 +128,5 @@ impl<'a> Debug for InformationRequestHeaderTagIter<'a> {
             debug.entry(e);
         });
         debug.finish()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{
-        HeaderTagFlag, InformationRequestHeaderTag, InformationRequestHeaderTagBuilder, MbiTagType,
-    };
-
-    #[test]
-    fn test_builder() {
-        let builder = InformationRequestHeaderTagBuilder::new(HeaderTagFlag::Required)
-            .add_ir(MbiTagType::EfiMmap)
-            .add_ir(MbiTagType::BootLoaderName)
-            .add_ir(MbiTagType::Cmdline);
-        // type(u16) + flags(u16) + size(u32) + 3 tags (u32)
-        assert_eq!(builder.expected_len(), 2 + 2 + 4 + 3 * 4);
-        let tag = builder.build();
-        let tag = unsafe {
-            (tag.as_ptr() as *const InformationRequestHeaderTag<3>)
-                .as_ref()
-                .unwrap()
-        };
-        assert_eq!(tag.flags, HeaderTagFlag::Required);
-        // type(u16) + flags(u16) + size(u32) + 3 tags (u32)
-        assert_eq!(tag.size, 2 + 2 + 4 + 3 * 4);
-        assert_eq!(tag.dynamic_requests_size(), 3);
-        assert!(tag.requests.contains(&MbiTagType::EfiMmap));
-        assert!(tag.requests.contains(&MbiTagType::BootLoaderName));
-        assert!(tag.requests.contains(&MbiTagType::Cmdline));
-        assert_eq!(tag.requests.len(), 3);
-        assert!(!tag.requests.contains(&MbiTagType::AcpiV1));
-        println!("{:#?}", tag);
     }
 }
