+++ multiboot2-header/src/lib.rs
@@ -3,8 +3,8 @@
 //!
 //! # Example
 //! ```rust
-//! use multiboot2_header::builder::Multiboot2HeaderBuilder;
-//! use multiboot2_header::{HeaderTagFlag, HeaderTagISA, InformationRequestHeaderTagBuilder, MbiTagType, RelocatableHeaderTag, RelocatableHeaderTagPreference, Multiboot2Header};
+//! use multiboot2_header::builder::{InformationRequestHeaderTagBuilder, Multiboot2HeaderBuilder};
+//! use multiboot2_header::{HeaderTagFlag, HeaderTagISA, MbiTagType, RelocatableHeaderTag, RelocatableHeaderTagPreference, Multiboot2Header};
 //!
 //! // Small example that creates a Multiboot2 header and parses it afterwards.
 //!
@@ -65,6 +65,9 @@ mod relocatable;
 mod tags;
 mod uefi_bs;
 
+#[cfg(feature = "builder")]
+pub mod builder;
+
 pub use self::address::*;
 pub use self::console::*;
 pub use self::end::*;
@@ -79,61 +82,6 @@ pub use self::relocatable::*;
 pub use self::tags::*;
 pub use self::uefi_bs::*;
 
-use core::mem::size_of;
 /// Re-export of [`multiboot2::TagType`] from `multiboot2`-crate as `MbiTagType`, i.e. tags that
 /// describe the entries in the Multiboot2 Information Structure (MBI).
 pub use multiboot2::TagType as MbiTagType;
-
-/// Trait for all tags that creates a byte array from the tag.
-/// Useful in builders to construct a byte vector that
-/// represents the Multiboot2 header with all its tags.
-#[cfg(feature = "builder")]
-pub(crate) trait StructAsBytes: Sized {
-    /// Returns the size in bytes of the struct, as known during compile
-    /// time. This doesn't use read the "size" field of tags.
-    fn byte_size(&self) -> usize {
-        size_of::<Self>()
-    }
-
-    /// Returns a byte pointer to the begin of the struct.
-    fn as_ptr(&self) -> *const u8 {
-        self as *const Self as *const u8
-    }
-
-    /// Returns the structure as a vector of its bytes.
-    /// The length is determined by [`size`].
-    fn struct_as_bytes(&self) -> alloc::vec::Vec<u8> {
-        let ptr = self.as_ptr();
-        let mut vec = alloc::vec::Vec::with_capacity(self.byte_size());
-        for i in 0..self.byte_size() {
-            vec.push(unsafe { *ptr.add(i) })
-        }
-        vec
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_as_bytes() {
-        struct Foobar {
-            a: u32,
-            b: u8,
-            c: u128,
-        }
-        impl StructAsBytes for Foobar {}
-        let foo = Foobar {
-            a: 11,
-            b: 22,
-            c: 33,
-        };
-        let bytes = foo.struct_as_bytes();
-        let foo_from_bytes = unsafe { (bytes.as_ptr() as *const Foobar).as_ref().unwrap() };
-        assert_eq!(bytes.len(), size_of::<Foobar>());
-        assert_eq!(foo.a, foo_from_bytes.a);
-        assert_eq!(foo.b, foo_from_bytes.b);
-        assert_eq!(foo.c, foo_from_bytes.c);
-    }
-}
